+++ ichnaea/app.py
@@ -1,15 +1,21 @@
 from pyramid.config import Configurator
 from pyramid.tweens import EXCVIEW
+from redis.exceptions import ConnectionError
+from sqlalchemy.exc import OperationalError
+from sqlalchemy.sql import func, select
 
 from ichnaea import customjson
 from ichnaea.cache import redis_client
-from ichnaea.db import Database
-from ichnaea.db import db_master_session
-from ichnaea.db import db_slave_session
+from ichnaea.db import (
+    Database,
+    db_master_session,
+    db_slave_session,
+    db_worker_session,
+)
 from ichnaea.geoip import configure_geoip
 
 
-def main(global_config, heka_config=None,
+def main(global_config, heka_config=None, init=False,
          _db_master=None, _db_slave=None, _heka_client=None, _redis=None,
          _stats_client=None, **settings):
     config = Configurator(settings=settings)
@@ -61,4 +67,24 @@ def main(global_config, heka_config=None,
 
     # replace json renderer with custom json variant
     config.add_renderer('json', customjson.Renderer())
+
+    # Should we try to initialize and establish the outbound connections?
+    if init:
+        # Test the slave DB connection
+        with db_worker_session(config.registry.db_slave) as session:
+            try:
+                session.execute(select([func.now()])).first()
+            except OperationalError:
+                # Let the instance start, so it can recover / reconnect
+                # to the DB later, but provide degraded service in the
+                # meantime.
+                pass
+
+        # Test the redis connection
+        try:
+            config.registry.redis_client.ping()
+        except ConnectionError:
+            # Same as for the DB, continue with degraded service.
+            pass
+
     return config.make_wsgi_app()